requirements.txt
tweepy==4.14.0
requests==2.31.0
python-dotenv==1.0.1
selenium==4.11.2
webdriver-manager==3.8.6
PySocks==1.7.1